third_party/blink/web_tests/external/wpt/editing/run/misc-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["defaultparagraphseparator",""]] "foo[bar]baz": execCommand("defaultparagraphseparator", false, "") return value
  assert_equals: expected false but got true
[FAIL] [["defaultparagraphseparator"," p "]] "foo[bar]baz": execCommand("defaultparagraphseparator", false, " p ") return value
  assert_equals: expected false but got true
[FAIL] [["defaultparagraphseparator","<p>"]] "foo[bar]baz": execCommand("defaultparagraphseparator", false, "<p>") return value
  assert_equals: expected false but got true
[FAIL] [["defaultparagraphseparator"," div "]] "foo[bar]baz": execCommand("defaultparagraphseparator", false, " div ") return value
  assert_equals: expected false but got true
[FAIL] [["defaultparagraphseparator","<div>"]] "foo[bar]baz": execCommand("defaultparagraphseparator", false, "<div>") return value
  assert_equals: expected false but got true
[FAIL] [["defaultparagraphseparator","li"]] "foo[bar]baz": execCommand("defaultparagraphseparator", false, "li") return value
  assert_equals: expected false but got true
[FAIL] [["defaultparagraphseparator","blockquote"]] "foo[bar]baz": execCommand("defaultparagraphseparator", false, "blockquote") return value
  assert_equals: expected false but got true
[FAIL] [["stylewithcss","true"]] "foo[bar]baz" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
Harness: the test ran to completion.
